refactor: migrate local_audio to party-display-core
