+++ app/src-tauri/party-display-core/src/lib.rs
@@ -2,3 +2,4 @@ pub mod dlna_proxy;
 pub mod system;
 pub mod smtc;
 pub mod dlna;
+pub mod local_audio;
